1. Added VLBA data support. 2. Added data modulation support, this is still    turned of by default for mark4 stations but still    is used in VLBA stations. Data modulation is    described in Mark4 memo 230A, whitney 2005, on    pages 6 and 7.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@896 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/data_reader.h
@@ -29,7 +29,7 @@ public:
       \return the number of bytes read into the buffer.
       \pre buff is allocated. nBytes >= 0
   **/
-  int get_bytes(size_t nBytes, char *buff);
+  size_t get_bytes(size_t nBytes, char *buff);
 
   /** Returns true if all data is read from the input reader.
   **/
@@ -69,7 +69,7 @@ public:
 private:
   /** Function that actually writes the data to the output device.
   **/
-  virtual int do_get_bytes(size_t nBytes, char *buff) = 0;
+  virtual size_t do_get_bytes(size_t nBytes, char *buff) = 0;
 
   uint64_t _data_counter;
   int data_slice;
