+++ include/data_reader_file.h
@@ -32,7 +32,7 @@ public:
   bool can_read();
 
 private:
-  int do_get_bytes(size_t nBytes, char *out);
+  size_t do_get_bytes(size_t nBytes, char *out);
 
   std::ifstream file;
 };
