+++ include/data_reader_socket.h
@@ -37,7 +37,7 @@ public:
   }
 
 protected:
-  int do_get_bytes(size_t nBytes, char *buff);
+  size_t do_get_bytes(size_t nBytes, char *buff);
 
   int m_socket;
   bool iseof;
